Suppress warning about narrowing type conversion in NCHW Convolution
PiperOrigin-RevId: 306721562

+++ src/convolution-nchw.c
@@ -650,7 +650,7 @@ static enum xnn_status setup_convolution2d_nchw(
     }
     case xnn_ukernel_type_dwconv:
     {
-      xnn_update_f32_spchw_params((union xnn_f32_spchw_params*) spchw_params, input_width);
+      xnn_update_f32_spchw_params((union xnn_f32_spchw_params*) spchw_params, (uint32_t) input_width);
       convolution_op->context.dwconv2d = (struct dwconv2d_context) {
         .output_height = output_height,
         .input_width = input_width,
